feat(mistral3): consolidate text decoder into MLXLMCommon namespace (WS-C #3, issue #168)
Lifts the Mistral 3 / Ministral 3 text-decoder layer stack into a
public `Mistral3` namespace under `MLXLMCommon/Models/Mistral3.swift`.
Both LLM (`Mistral3TextModel`, renamed file from Mistral3Text.swift)
and VLM (`Language.LanguageModel`) consume the shared types.

The two configs (`Mistral3TextConfiguration` LLM-side,
`Mistral3VLMTextConfiguration` VLM-side) use different field names
(attentionHeads vs numAttentionHeads, vocabularySize vs vocabSize).
Rather than force-unify the configs, the namespace's layer classes
take a thin `Mistral3.LayerArgs` adapter struct that each config
produces via a computed `var layerArgs` property.

Shared:
- `Mistral3.LayerArgs` (adapter)
- `Mistral3.llama4AttentionScale(start:stop:beta:maxPositionEmbeddings:)`
- `Mistral3.Attention` (uses applyRotaryPosition; takes attnScale param
  for Llama-4 position-based scaling)
- `Mistral3.MLP`
- `Mistral3.TransformerBlock` (carries useSliding flag)
- `Mistral3.ModelInner` (full-attention vs sliding-attention dispatch
  via layerTypes; computes Llama-4 scaling internally)

Per-target wrappers:
- `MLXLLM/Models/Mistral3.swift` (renamed from Mistral3Text.swift):
  `Mistral3TextModel: LLMModel` + Configuration with dual-decode
  (flat + nested `text_config` JSON paths). Optional Linear lm_head
  (or tied embedding). `weight_scale_inv` legacy quant handling
  preserved.
- `MLXVLM/Models/Mistral3.swift`: `Language.LanguageModel` keeps its
  `Linear?` lm_head and per-layer cache strategy. Adds preemptive
  prefill-sync `eval()` barrier to `Mistral3VLM.prepare` mirroring the
  Gemma 3 / Gemma 4 VLM fix from PR #172.

Verification:
- swift build clean
- swift test --configuration release: 203/204 (1 pre-existing
  benchmark scaffold failure)
- LLM smoke `mlx-community/Ministral-3-3B-Instruct-2512-4bit
  --method simple`: "I'm ChatGPT — your AI assistant..." (93 tok/s)
- VLM smoke deferred (no cached Mistral3 VLM checkpoint locally).
